+++ manifests/common/crd/odhdashboardconfigs.opendatahub.io.crd.yaml
@@ -213,6 +213,8 @@ spec:
                       message: 'DEPRECATED: spec.dashboardConfig.disableFineTuning must be removed or left unchanged.'
                     - rule: '!has(self.mlflow) || (has(oldSelf.mlflow) && self.mlflow == oldSelf.mlflow)'
                       message: 'DEPRECATED: spec.dashboardConfig.mlflow must be removed or left unchanged.'
+                    - rule: '!has(self.maasAuthPolicies) || self.maasAuthPolicies == oldSelf.maasAuthPolicies'
+                      message: 'DEPRECATED: spec.dashboardConfig.maasAuthPolicies must be removed or left unchanged.'
                 # TODO: Remove before going to v1
                 groupsConfig:
                   description: 'Ignored -- See "Auth" Resource'
