+++ packages/cypress/cypress/tests/mocked/modelsAsAService/maasExternalModels.cy.ts
@@ -1,8 +1,11 @@
 import { mockDashboardConfig } from '@odh-dashboard/internal/__mocks__/mockDashboardConfig.js';
 import { mockDscStatus } from '@odh-dashboard/internal/__mocks__/mockDscStatus.js';
 import { DataScienceStackComponent } from '@odh-dashboard/plugin-core/areas';
-import { externalModelsPage } from '../../../pages/modelsAsAService';
+import { ProjectModel } from '@odh-dashboard/internal/api/models/index';
+import { mockK8sResourceList } from '@odh-dashboard/internal/__mocks__/mockK8sResourceList';
+import { mockProjectK8sResource } from '@odh-dashboard/internal/__mocks__/mockProjectK8sResource';
 import { asProductAdminUser } from '../../../utils/mockUsers';
+import { externalModelsPage } from '../../../pages/modelsAsAService';
 
 const setupCommonIntercepts = () => {
   asProductAdminUser();
@@ -13,6 +16,7 @@ const setupCommonIntercepts = () => {
   cy.interceptOdh('GET /maas/api/v1/user', {
     data: { userId: 'test-user', clusterAdmin: false },
   });
+  cy.interceptK8sList(ProjectModel, mockK8sResourceList([mockProjectK8sResource({})]));
   cy.interceptOdh('GET /maas/api/v1/namespaces', { data: [] });
   cy.interceptOdh(
     'GET /api/dsc/status',
@@ -36,7 +40,10 @@ describe('External Models Page', () => {
     externalModelsPage.findDescription().should('exist');
   });
   it('should not show the external models page when the feature flag is disabled', () => {
-    cy.interceptOdh('GET /api/config', mockDashboardConfig({ modelAsService: true }));
+    cy.interceptOdh(
+      'GET /api/config',
+      mockDashboardConfig({ modelAsService: true, externalModels: false }),
+    );
     externalModelsPage.visit();
     externalModelsPage.findExternalModelsTab().should('not.exist');
     externalModelsPage.findTitle().should('not.exist');
